refactor(driver): remove the legacy InsertSpec compatibility path
With every workload now loading through the typed driver.Insert /
bench.Bench.Insert path over gen.BatchSource, the InsertSpec
compatibility layer is dead. Remove it.

Gone:
- Driver.InsertSpec (interface), Bench.InsertSpec +
  newInsertProgressTracker + insertMethodName (pkg/bench/query.go).
- Per-driver InsertSpec methods (postgres, mysql, picodata, ydb, csv,
  noop) and sqldriver.RunInsertSpec; the typed Insert method,
  runInsertChunk, and the COPY/bulk/columnar helpers stay.
- pkg/datagen/loadsource (no remaining callers).
- common.RunParallelByWorkers + RunParallel (legacy
  source.Partitionable runners); the shared Chunk type + SplitChunks
  move to chunks.go, reused by the typed RunParallelBatch.
- driver.MethodFromProto / MethodToProto (dgproto<->driver boundary
  converters); the InsertMethod enum, ParseInsertMethod, String, and
  capability matrix stay.
- dispatcher.ErrInsertSpecNotImplemented (unused).
- Legacy-compat tests: sqldriver/insert_spec_test, noop/driver_test,
  common/parallel_insert_test (SplitChunks test moved to chunks_test),
  integration/smoke_driver_insert_spec. csv/driver_test trimmed to its
  run-query/config/manifest tests; TestManifestWritten migrated to
  the typed Insert path. insert_method_test trimmed of the proto
  roundtrip.

No production caller of InsertSpec remained after the tpch/tpcds
ports; noop loads confirm unchanged cardinalities (tpcc 569011,
tpch 86630, tpcds 2658555). Build, full suite, race, and linter
green.
