+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Group lines under `Added` / `Changed` / `Fixed` / `Removed`. Append a PR link
 - TPC-C complex tables (`customer`, `stock`, `orders`, `order_line`) load through the typed insert path: the per-district `o_c_id` Feistel permutation, NURand `c_last`, the ~10% `BC`/`GC` credit split, the 10 `s_dist_NN` columns, the ORIGINAL `s_data` marker, the delivered/undelivered `o_carrier_id`/`ol_delivery_d`/`ol_amount` split, and the first-1000 surname mapping are plain Go over bound `gen` columns. The TPC-C expression/protobuf helpers are gone.
 - TPC-H loads through the typed insert path: all eight tables (`region`, `nation`, `part`, `supplier`, `partsupp`, `customer`, `orders`, `lineitem`) stream from a `gen.BatchSource` adapter over the canonical dbgen generator, so `InsertTpch` no longer synthesizes a protobuf `InsertSpec`. Canonical dbgen seeds, per-district seeking, entity fan-out, and SF=1 output are unchanged (the FNV golden hashes still match). The dbgen `Make*` layer still allocates `[]any` per entity internally — that is the documented allocation boundary, not the typed adapter.
 - TPC-DS loads through the typed insert path: all 24 tables (18 dimension tables, inventory, and 6 fan-out sales/returns fact tables) stream from a `gen.BatchSource` adapter over the canonical dsdgen generator, so `InsertTpcds` no longer synthesizes a protobuf `InsertSpec`. Canonical dsdgen text output, null semantics, ticket fan-out, and nominal fact-row reporting are unchanged.
+- The legacy `InsertSpec` load path is gone. `Driver.InsertSpec`, `Bench.InsertSpec`, `loadsource.Build`, the per-driver `InsertSpec` methods, `RunParallelByWorkers`, and the dgproto↔driver `MethodFromProto`/`MethodToProto` boundary converters are removed; every workload now loads exclusively through the typed `driver.Insert`/`Bench.Insert` path over `gen.BatchSource`. The shared `Chunk`/`SplitChunks` helpers moved to `pkg/driver/common/chunks.go`; per-driver `runInsertChunk`/bulk/COPY helpers are unchanged. Insert-method strings (`plain_query`, `plain_bulk`, `columnar`, `native`), probe output, and progress/metrics semantics are preserved.
 - Metrics can again be exported to an OpenTelemetry collector through the existing `global.exporter.otlpExport` gRPC or HTTP configuration. ([#125](https://github.com/stroppy-io/stroppy/pull/125))
 
 ### Changed
