+++ pkg/driver/dispatcher.go
@@ -9,7 +9,6 @@ import (
 	"go.uber.org/zap"
 
 	stroppy "github.com/stroppy-io/stroppy/pkg/common/proto/stroppy"
-	"github.com/stroppy-io/stroppy/pkg/datagen/dgproto"
 	"github.com/stroppy-io/stroppy/pkg/driver/stats"
 )
 
@@ -45,14 +44,10 @@ type (
 	}
 
 	Driver interface {
-		// InsertSpec runs a relational InsertSpec through the driver, streaming
-		// rows from a dgproto-driven runtime.Runtime into the database.
-		InsertSpec(ctx context.Context, spec *dgproto.InsertSpec) (*stats.Query, error)
 		// Insert runs a typed [InsertRequest] through the driver, streaming
 		// rows from a workload-authored [gen.BatchSource] into the database.
 		// Generation after cursor preparation allocates nothing; driver-side
-		// encoding and materialization may. It is the typed successor to
-		// InsertSpec and coexists with it until the legacy path is removed.
+		// encoding and materialization may.
 		Insert(ctx context.Context, req *InsertRequest) (*stats.Query, error)
 		RunQuery(ctx context.Context, sql string, args map[string]any) (*QueryResult, error)
 		Begin(ctx context.Context, isolation stroppy.TxIsolationLevel) (Tx, error)
@@ -63,12 +58,6 @@ type (
 )
 
 var ErrNoRegisteredDriver = errors.New("no registered driver")
-
-// ErrInsertSpecNotImplemented is returned by drivers that have not yet
-// wired up the relational InsertSpec path. Drivers opt in by implementing
-// the full InsertSpec method; until then they return this sentinel so
-// callers can distinguish "not wired" from a genuine driver error.
-var ErrInsertSpecNotImplemented = errors.New("driver: InsertSpec not implemented")
 
 var registry = map[stroppy.DriverConfig_DriverType]driverConstructor{}
